Reusable MockStore for Storybook
